+++ src/math/Ray.java
@@ -15,23 +15,23 @@ import utils.GeometryUtil;
  */
 public class Ray {
 
-    public double x1,y1,x2,y2;
-    public Point2D.Double source;
-    public Ray(double x1, double y1, double x2, double y2) {
+    public float x1,y1,x2,y2;
+    public Point2D.Float source;
+    public Ray(float x1, float y1, float x2, float y2) {
         this.x1 = x1;
         this.y1 = y1;
         this.x2 = x2;
         this.y2 = y2;
-        source = new Point2D.Double(x1,y1);
+        source = new Point2D.Float(x1,y1);
     }
     
     
-    public Ray(Point2D.Double from, Point2D.Double to){
+    public Ray(Point2D.Float from, Point2D.Float to){
         x1 = from.x;
         y1 = from.y;
         x2 = to.x;
         y2 = to.y;
-        source = new Point2D.Double(x1,y1);
+        source = new Point2D.Float(x1,y1);
     }
